EFB Maps: prevent concurrency artifacts
Previously, zooming or scrolling the map would sometimes cause incorrect
tiles to be shown, due to concurrency issues between multiple HTTP
requests. We would also keep downloading tiles even when it is already
clear that we won't need them.

We also limit the zoom and scroll ranges, and handle out-of-bounds tile
indices more gracefully: on the X axis (latitude), we wrap around,
providing seamless East-West scrolling in either direction; on the Y
axis (longitude), we limit the map scroll position to 77.5° N and S, and
hide any out-of-bound tiles.